remove unused structs
